Start updating confidential asset tests.
